New twitter client added
Now twitter scrapper is injected into the main application strucure. Two implementations already created. We call them "go-twitter" and "void".
