+++ cmd/server/server.go
@@ -1,33 +1,51 @@
 package main
 
 import (
+	"errors"
+	"fmt"
 	"strings"
 
 	"github.com/dairlair/tweetwatch/pkg/cmd/server"
 	grpcServer "github.com/dairlair/tweetwatch/pkg/protocol/grpc"
 	"github.com/dairlair/tweetwatch/pkg/storage"
 	"github.com/dairlair/tweetwatch/pkg/twitterclient"
+	"github.com/dairlair/tweetwatch/pkg/twitterclient/providers/gotwitter"
+	"github.com/dairlair/tweetwatch/pkg/twitterclient/providers/void"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/viper"
 )
 
 func main() {
-	config := readConfig()
-	log.Infof("Config: %v\n", config)
-
-	srv := server.NewInstance(&config)
-	err := srv.Start()
+	config, providers, err := readConfig()
 	if err != nil {
-		log.Errorf("twitwatch start failed: %s", err)
+		log.Errorf("tweetwatch configurations failed: %s", err)
+		return
+	}
+	log.Infof("config: %v\n", config)
+
+	srv := server.NewInstance(&config, providers)
+	err = srv.Start()
+	if err != nil {
+		log.Errorf("tweetwatch start failed: %s", err)
 	}
 }
 
-func readConfig() server.Config {
+func readConfig() (server.Config, server.Providers, error) {
 	configureViper()
 
 	err := viper.ReadInConfig()
 	if err != nil {
 		log.Warnf("config file not read: %s", err)
+	}
+
+	twitterProviderName := viper.GetString("twitter.provider")
+	twitterProvider, err := getTwitterProvider(twitterProviderName)
+	if err != nil {
+		return server.Config{}, server.Providers{}, err
+	}
+
+	providers := server.Providers{
+		CreateTwitterclient: twitterProvider,
 	}
 
 	return server.Config{
@@ -43,7 +61,7 @@ func readConfig() server.Config {
 			TwitterAccessToken:    viper.GetString("twitter.accessToken"),
 			TwitterAccessSecret:   viper.GetString("twitter.accessSecret"),
 		},
-	}
+	}, providers, nil
 }
 
 func configureViper() {
@@ -55,4 +73,18 @@ func configureViper() {
 	viper.AddConfigPath("./")
 	viper.SetEnvKeyReplacer(strings.NewReplacer(".", "_"))
 	viper.SetDefault("grpc.listen", ":1308")
+	viper.SetDefault("twitter.provider", "go-twitter")
+}
+
+func getTwitterProvider(provider string) (server.TwitterClientProvider, error) {
+	log.Infof("twitter provider: %s", provider)
+	switch provider {
+	case "void":
+		return void.NewInstance, nil
+	case "go-twitter":
+		return gotwitter.NewInstance, nil
+	}
+
+	msg := fmt.Sprintf("unknown twitter provider [%s]. available values: \"go-twitter\", \"void\"\n", provider)
+	return nil, errors.New(msg)
 }
