+++ pkg/twitterclient/providers/gotwitter/README.md
@@ -1,0 +1,4 @@
+# gotwitter package
+
+This package provides structure that implements twitterclient.Interface 
+through [go-twitter](https://github.com/dghubble/go-twitter) library.
