+++ pkg/twitterclient/providers/void/README.md
@@ -1,0 +1,5 @@
+# Void package
+
+This package provides structure that implements twitterclient.Interface 
+without any data retrieving or network connections. You can use this provider 
+through configuration to avoid undesired network traffic consumption.
