--- pkg/twitterclient/stream.go
@@ -1,1 +1,0 @@
-package twitterclient
